+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # 0.47.0
 
 -   Breaking Change: Default to dynamic object policy in `convertSchema`. See docs on `strictMode` option.
--   Added `failedRecords` to `process` event.
+-   Added `failedRecords` to `process` event and `failedRecord` to `error` for changestream events where applicable.
 -   Added `options.mapper` to `initSync`.
 
 # 0.46.0